Move matrix construction Ops to sub-module
